+++ games/static/js/add_game.js
@@ -1,0 +1,12 @@
+import { syncSelectInputUntilChanged } from './utils.js'
+
+let syncData = [
+  {
+    "source": "#id_name",
+    "source_value": "value",
+    "target": "#id_sort_name",
+    "target_value": "value"
+  }
+]
+
+syncSelectInputUntilChanged(syncData, "form")